Make --region argument mandatory for AWS related commands.
While some AWS services may work without specifying a region (that's
why we make this argument optional at the beginning), others need
it. Instead of letting the end user figure out in which cases it's
needed and in which cases is not, we've decided to make it mandatory
in all cases (and reduce the possible confusion).