Rename profiles to better reflect that they are about deployments
Previously we only printed the names of the deployment targets. But we
didn't reflect the fact that those names were about where to deploy
things. Which as a bit misleading in the case of the "Amazon Web
Services" (old) profile name. Because users sometimes interpreted that
profile name as "I want to use AWS services like S3 or Cognito" (even
if the user wanted to deploy her project on-premises). And not as "I
want to deploy my project in AWS".