Add restrictions on project names
We now impose a limit on the size of the project name. The project
name maximum length is now 48 characters. The reason for this it that
quite a few AWS resource names have a limit of 64 characters. And we
use the project name, with some extra suffixes, to build those AWS
resource names. We are reserving 16 characters for those suffixes, so
that limits the project name to 48 characters max.

On the other hand, we now allow numbers in the project names, but not
in the first position. Project names are used in a few places (like
YAML files, or CloudFormation templates) where values cannot start
with a number.