Use shorter suffixes for IAM role names
As mentioned in commit beb51a2872, many AWS resource names have a
maximum length of 64 characters. We limited the project name length in
that commit to 48 characters, to leave room for the suffixes we use to
name certain AWS resources (16 characters).

In the case of these two role resources, the suffixes we used went
over the 16 characters limit. So we are changing them to stay within
the limit.