Stop using reitit.dev.pretty/exception
Even if it's part of reitit-dev, we were using it in production too.

We have decided that it was not necessary, and that it was just
a (small) convenience.

Thus we have removed it and will use the default. In fact, we won't
get those exceptions in "production" (release mode). Those exceptions
are only thrown when compiling the routes. And we would get them in
local development first, and would fix them there. Otherwise the
router wouldn't even start, and the backend (or front end) routes
wouldn't even work at all.

[Re: #29]