Add missing mandatory flag to predeploy script
The HOP CLI command to download SSM parameter store variables is
missing the mandatory --region flag. In order to determine what region
needs to be used (because it depends on where the ElasticBeanstalk is
deployed) we query the EC2 instance meta-data service.

The assumption here is that ElasticBeanstalk will create the EC2
instances it uses in the same region as the environment that is
creating the instances.

[Re: #35]