Add JVM option to REPL profile, to be able to stop runaway threads.
Since version 1.2.0, nREPL includes a native JVMTI agent which makes
the eval interrupts work properly on Java 21 and later. To enable the
agent, the Java process should be launched with
-Djdk.attach.allowAttachSelf.

While some development tooling, like CIDER, may set this option by
itself if launching the REPL in a certain way, this doesn't work in
the case of starting the REPL outside the development tooling and then
connecting to it. In that scenario, you have to add
-Djdk.attach.allowAttachSelf Java property manually through the means
of the build tool.