Set execution permissions Babashka scripts too
We are only setting execution permissions for files with `sh`
extension. Thus, making Babashka scripts such as
merge-docker-compose-files.clj non executable from the get go, causing
unexpected CI/CD failure.

We don't want to add the `clj` extension to the condition check,
because in that case we would be making all the other clj files
executable. And we don't want that.

One possible solution is to use the `bb` extension on the Babashka
scripts, which is also a supported extension. We think this is a good
approach, because this way we differentiate the JVM Clojure files from
Babashka scripts in the code base. And this differentiation makes
setting the execution permissions quite easy.

[Re: #34]