Move api/config files to the frontend profile.
That API endpoing it actually only used by the frontend profile. So it
makes more sense to put it there. That way, if the project doesn't
contain the frontend profile, the API endpoit is not added.

Also, prepare the route metadata for the upcoming migration from
Swagger 2.x to OpenAPI 3.x